Últimos cambios espero. Transfiriendo la lógica del negocio a la capa de servicios. Arreglando un error en la búsqueda de correos. Diagrama UML actualizado.

+++ src/main/java/app/servicios/ServiciosCliente.java
@@ -1,10 +1,11 @@
 package app.servicios;
 
 import java.util.List;
-import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 import app.entidades.Cliente06;
 import app.repositorios.ClientesRepositorio;
@@ -20,27 +21,50 @@ public class ServiciosCliente implements RequerimientosCRUD<Cliente06> {
 	private ClientesRepositorio clientesRepositorio;
 	@Override
 	public List<Cliente06> listarTodos() {
-        return clientesRepositorio.findAll();
+		List<Cliente06> lista = clientesRepositorio.findAll();
+		if (lista == null || lista.isEmpty() || lista.size() == 0) {
+			return List.of();
+		}
+		return lista; // Reemplaza con la lista de clientes obtenida
     }
 
-	public List<Cliente06> search(String searchTermn){
-		return clientesRepositorio.findByDniContainingOrNombreContainingOrApellidoAllIgnoreCase(searchTermn, searchTermn, searchTermn);
+	public List<Cliente06> search(String searchTerm){
+		searchTerm = searchTerm.trim();
+		if (searchTerm.isBlank()){
+			return this.listarTodos();
+		}
+		return clientesRepositorio.findByDniContainingOrNombreContainingOrApellidoContainingAllIgnoreCase(searchTerm, searchTerm, searchTerm);
 	}
 
 	@Override
 	public void actualizar(Cliente06 cliente) {
-		clientesRepositorio.save(cliente);
+		if (clientesRepositorio.existsById(cliente.getDni())) {
+			clientesRepositorio.save(cliente);
+		} else {
+			throw new ResponseStatusException( HttpStatus.NOT_FOUND, "Cliente no encontrado con DNI: " + cliente.getDni() + ". No se pudo actualizar.");
+		}
 	}
 	@Override
 	public Cliente06 buscarPorId(String dni) {
+		if (!clientesRepositorio.existsById(dni)) {
+			return new Cliente06();
+		}
         return clientesRepositorio.findById(dni).orElse(null);
     }
 	@Override
 	public void guardar(Cliente06 cliente) {
-		clientesRepositorio.save(cliente);
+		if (clientesRepositorio.existsById(cliente.getDni())) {
+			clientesRepositorio.save(cliente);
+		} else {
+			throw new ResponseStatusException(HttpStatus.CONFLICT, "Existe un cliente con ese DNI: " + cliente.getDni() + ". No se ha agregado un Cliente nuevo.");
+		}
+
     }
 	@Override
 	public void eliminarPorId(String dni) {
+		if (!clientesRepositorio.existsById(dni)) {
+			throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Cliente no encontrado con DNI: " + dni + ". No se pudo eliminar.");
+		}
 		clientesRepositorio.deleteById(dni);
     }
 	@Override
